Move scheduler.py to services folder
