--- garnbarn_api/management/commands/runscheduler.py
@@ -1,29 +1,0 @@
-# runapscheduler.py
-import logging
-
-from django.conf import settings
-
-from apscheduler.schedulers.blocking import BlockingScheduler
-from apscheduler.triggers.cron import CronTrigger
-from django.core.management.base import BaseCommand
-from django_apscheduler.jobstores import DjangoJobStore
-from django_apscheduler.models import DjangoJobExecution
-from django_apscheduler import util
-
-logger = logging.getLogger(__name__)
-
-scheduler = BlockingScheduler(timezone=settings.TIME_ZONE)
-scheduler.add_jobstore(DjangoJobStore(), "default")
-
-
-class Command(BaseCommand):
-    help = "Runs APScheduler."
-
-    def handle(self, *args, **options):
-        try:
-            logger.info("Starting scheduler...")
-            scheduler.start()
-        except KeyboardInterrupt:
-            logger.info("Stopping scheduler...")
-            scheduler.shutdown()
-            logger.info("Scheduler shut down successfully!")
