+++ garnbarn_api/models.py
@@ -6,9 +6,8 @@ import math
 from datetime import datetime, timedelta
 from garnbarn_api.services.pubsub import pubsub
 from rest_framework import serializers
-from garnbarn_api.scheduler import scheduler
+from garnbarn_api.services.scheduler import scheduler
 from apscheduler.triggers.date import DateTrigger
-from scheduler import scheduler
 
 
 class CustomUser(models.Model):
@@ -111,10 +110,6 @@ class Assignment(models.Model):
         """
         super().save(*args, **kwargs)
         self.refresh_from_db()
-        # TODO: For development purpose This scheduler will schedule the job for the next 10 seconds after the call time.
         schedule_date = datetime.now() + timedelta(seconds=10)
         scheduler.add_job(pubsub, trigger=DateTrigger(run_date=schedule_date), id=f"Notification - {self.pk}",
                           max_instances=1, replace_existing=True)
-        # TODO: Remove these line.
-        # print(a.trigger)
-        # print(scheduler.get_jobs())
